+++ docs/conf.py
@@ -1,6 +1,6 @@
 project = "mechai-model-selection"
-author = "Pengfei Song and Yin Song"
-release = "0.3.0"
+author = "Shuangshuang Yin and Pengfei Song"
+release = "0.3.1"
 extensions = ["sphinx.ext.autodoc", "sphinx.ext.napoleon", "sphinx.ext.viewcode"]
 html_theme = "sphinx_rtd_theme"
 autodoc_typehints = "description"